requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.10
django-allauth==65.13.1
django-filter==25.2
django-user-agents==0.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-auth-kit==1.1.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
structlog==25.5.0
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==202601
uritemplate==4.2.0
user-agents==2.2.0
